SVN revision 12882

+++ src/proto/endecode-funcs.h
@@ -1577,20 +1577,20 @@ static inline void defree_##name(struct name *x) {                            \
 
 /* 3 fields, then an array, then 2 fields, then an array */
 #define endecode_fields_3a2a_struct(name, t1, x1, t2, x2, t3, x3, tn1, n1, ta1, a1, t4, x4, t5, x5, tn2, n2, ta2, a2) \
-static inline void encode_##name(char **pptr, const struct name *x) { int i; \
+static inline void encode_##name(char **pptr, const struct name *x) { int i;  \
     encode_##t1(pptr, &x->x1); \
     encode_##t2(pptr, &x->x2); \
     encode_##t3(pptr, &x->x3); \
     encode_##tn1(pptr, &x->n1); \
     if (x->n1 > 0) \
-        for (i=0; i<x->n1; i++) \
+        for (i=0; i<(int)(x->n1); i++) \
             encode_##ta1(pptr, &(x)->a1[i]); \
     align8(pptr); \
     encode_##t4(pptr, &x->x4); \
     encode_##t5(pptr, &x->x5); \
     encode_##tn2(pptr, &x->n2); \
     if (x->n2 > 0) \
-        for (i=0; i<x->n2; i++) \
+        for (i=0; i<(int)(x->n2); i++) \
             encode_##ta2(pptr, &(x)->a2[i]); \
     align8(pptr); \
 } \
@@ -1602,7 +1602,7 @@ static inline void decode_##name(char **pptr, struct name *x) { int i; \
     if (x->n1 > 0) \
     { \
         x->a1 = decode_malloc(x->n1 * sizeof(*x->a1)); \
-        for (i=0; i<x->n1; i++) \
+        for (i=0; i<(int)(x->n1); i++) \
             decode_##ta1(pptr, &(x)->a1[i]); \
     } \
     else \
@@ -1614,7 +1614,7 @@ static inline void decode_##name(char **pptr, struct name *x) { int i; \
     if (x->n2 > 0) \
     { \
         x->a2 = decode_malloc(x->n2 * sizeof(*x->a2)); \
-        for (i=0; i<x->n2; i++) \
+        for (i=0; i<(int)(x->n2); i++) \
             decode_##ta2(pptr, &(x)->a2[i]); \
     } \
     else \
@@ -1638,14 +1638,14 @@ static inline void encode_##name(char **pptr, const struct name *x) { int i; \
     encode_##t3(pptr, &x->x3); \
     encode_##tn1(pptr, &x->n1); \
     if (x->n1 > 0) \
-        for (i=0; i<x->n1; i++) \
+        for (i=0; i<(int)(x->n1); i++) \
             encode_##ta1(pptr, &(x)->a1[i]); \
     align8(pptr); \
     encode_##t4(pptr, &x->x4); \
     encode_##t5(pptr, &x->x5); \
     encode_##tn2(pptr, &x->n2); \
     if (x->n2 > 0) \
-        for (i=0; i<x->n2; i++) \
+        for (i=0; i<(int)(x->n2); i++) \
             encode_##ta2(pptr, &(x)->a2[i]); \
     align8(pptr); \
     encode_##t6(pptr, &x->x6); \
@@ -1659,7 +1659,7 @@ static inline void decode_##name(char **pptr, struct name *x) { int i; \
     if (x->n1 > 0) \
     { \
         x->a1 = decode_malloc(x->n1 * sizeof(*x->a1)); \
-        for (i=0; i<x->n1; i++) \
+        for (i=0; i<(int)(x->n1); i++) \
             decode_##ta1(pptr, &(x)->a1[i]); \
     } \
     else \
@@ -1671,7 +1671,7 @@ static inline void decode_##name(char **pptr, struct name *x) { int i; \
     if (x->n2 > 0) \
     { \
         x->a2 = decode_malloc(x->n2 * sizeof(*x->a2)); \
-        for (i=0; i<x->n2; i++) \
+        for (i=0; i<(int)(x->n2); i++) \
             decode_##ta2(pptr, &(x)->a2[i]); \
     } \
     else \
